+++ app/src/main/java/com/donnKey/aesopPlayer/ui/provisioning/ProvisioningActivity.java
@@ -204,10 +204,7 @@ public class ProvisioningActivity extends AppCompatActivity
             setMenuItemProperties(this, item,
                     enabled ? R.drawable.ic_settings_red_24dp : R.drawable.ic_settings_redish_24dp,
                     enabled ? android.R.color.white : R.color.medium_dark_grey);
-            globalSettings.setMaintenanceMode(enabled);
-            if (!enabled) {
-                KioskModeSwitcher.enableAccessibilityIfNeeded(this);
-            }
+            KioskModeSwitcher.enableMaintenanceMode(this, enabled);
             return true;
         }
         return false;